Move forms to projects/forms.py
